Исправил типы данных

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
         int weight = 65;
-        double height = 174;
-        double index = weight / (height * height / 10_000);
+        int height = 174;
+        int index = weight / (height * height / 10_000);
         System.out.println(index);
     }
 }
